+++ front/src/types/settingsModel.ts
